move App.tsx to ./app and create HOC error-bounder
