+++ src/index.tsx
@@ -1,4 +1,7 @@
 import ReactDOM from 'react-dom';
-import { App } from './App';
+import { withErrorBoundary } from './hocs/error-boundary';
+import { App } from './app';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const AppWithErrorBoundary = withErrorBoundary(App);
+
+ReactDOM.render(<AppWithErrorBoundary />, document.getElementById('root'));
